+++ tests/no_features_errors/from/short_unfinished_identifier_list.rs
@@ -1,0 +1,6 @@
+use duplicate::*;
+#[duplicate_item(
+	ident1 ident2
+)]//duplicate_end
+pub struct name(ty);
+//item_end