+++ tests/pretty_errors/expected/only_semicolon.expanded.rs
@@ -1,0 +1,1 @@
+substitution_identifier (short syntax) or substitution group (verbose syntax)