+++ tests/pretty_errors/expected/parameters_not_encapsulated.expanded.rs
@@ -1,5 +1,4 @@
-Expected '['.
-       Substitution parameters should be enclosed in '[]' each.
+       Hint: Substitution parameters should be enclosed in '[]' each.
        Example:
            sub_ident( [ parameter1 ] , [ paramter2 ] )
                      ^^^          ^^^ ^^^         ^^^