+++ tests/pretty_errors/expected/short_unfinished_identifier_list.expanded.rs
@@ -1,0 +1,1 @@
+substitution_identifier or ';'