+++ tests/pretty_errors/expected/short_without_duplicates.expanded.rs
@@ -1,0 +1,5 @@
+       Hint: Add a substitution group after the substitution identifiers.
+       Example:
+           name;
+           [SomeSubstitution];
+           ^^^^^^^^^^^^^^^^^^^