+++ tests/pretty_errors/from/no_inputs.rs
@@ -1,0 +1,6 @@
+use duplicate::*;
+#[duplicate_item(
+
+)]//duplicate_end
+pub struct name(ty);
+//item_end