+++ tests/pretty_errors/from/short_without_duplicates.rs
@@ -1,0 +1,7 @@
+use duplicate::*;
+#[duplicate_item(
+	ty 		[i16];
+	name;
+)]//duplicate_end
+pub struct name(ty);
+//item_end